requirements.txt
asgiref==3.7.2
Django==4.2.5
django-froala-editor==4.1.2
Pillow==10.0.0
psycopg2-binary
python-dotenv==1.0.0
sqlparse==0.4.4
typing_extensions==4.7.1
